add stderr from external processes fix CLI class
