+++ src/main/java/com/galoev/CommandEcho.java
@@ -1,3 +1,5 @@
+package com.galoev;
+
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.util.List;
