+++ src/main/java/com/galoev/CommandPwd.java
@@ -1,3 +1,5 @@
+package com.galoev;
+
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.nio.file.Paths;
