+++ src/main/java/com/galoev/Substitute.java
@@ -1,3 +1,5 @@
+package com.galoev;
+
 import java.util.ArrayList;
 import java.util.List;
